# crates/cosmon-rpp-adapter/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Cosmon Remote Pilot Port (RPP) — V0+V1
  version: "1.0.0"
  description: |
    Hand-curated tenant client contract for the Remote Pilot Port. Requests
    execute through the admission boundary and injectable library ports.
    Worker dispatch may spawn processes; tenant state remains isolated.

    The registered surface is partitioned between this document and exact
    entries in openapi/exclusions.json, each with a reviewed reason. The
    cargo integration test openapi_surface checks both directions of path
    coverage directly against src/lib.rs, independently of the event canon.
    This file is not generated from Rust types. Path coverage does not prove
    method, schema or behavioral parity; handler changes still require review.

    Bounded run (ADR-124) and sealed done (ADR-176 as amended) are tenant
    request doors. Their presence does not expose the whole CLI. Operator
    provisioning, bootstrap distribution and the separate MCP protocol are
    outside this published REST contract. Existing healthz, metrics and
    diagnostics remain documented operational endpoints without tenant data.
  contact:
    name: Noogram
    url: https://github.com/noogram/cosmon
  license:
    name: GNU Affero General Public License v3.0 only
    identifier: AGPL-3.0-only
servers:
  - url: https://rpp.{domain}
    description: Operator-controlled deployment (typical tenant_auditor-facing instance).
    variables:
      domain:
...
security:
  - bearerAuth: []
paths:
...
  /v1/vitals:
    get:
      operationId: getVitals
      summary: Read observed fleet health for the caller's tenant.
      description: |
        Returns one row per non-terminal molecule in the caller's noyau.
        Durable lifecycle `status` and observed transport `health` are
        separate fields: a recorded process that no longer exists is
        `orphaned`, while a confirmed session is `live`. Probe failures are
        `unknown`; molecules without a process record are `unassigned`.

        Counts, rows, operator-wait declarations, and backlog age are folded
        once by `cosmon_state::ops::vitals`. Backlog staleness uses the same
        `cosmon_core::staleness` threshold as `cs status` and `cs peek`.

        This is an adapter-only tenant view, not the operator-wide `cs peek`
        raster. It does not inspect panes or transcripts for mute hangs.

        ## Required scope

        `cosmon:molecule:read` OR `cosmon:molecule:write` (write implies
        visibility).
      tags: [molecules]
      responses:
        "200":
          description: Tenant fleet vitals (possibly with no rows).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalsEnvelope"
        "401":
          description: Missing, malformed, or rejected JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: JWT valid but lacks molecule read scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: The bound tenant workspace does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Per-`sub` rate limit exhausted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: State store unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /v1/workers:
    get:
      summary: |
        List active workers in the per-tenant noyau (gap report
        ae3d workflow §e priority 5, `task-20260523-f82b`).
      description: |
        Return the set of workers currently bound to molecules in the
        caller's noyau. The route is read-only and tenant-pinned —
        a noyau-A JWT only ever sees noyau-A workers; cross-tenant
        enumeration is structurally impossible.

        ## Source of truth

        The handler scans the per-tenant store for molecules carrying
        a live [`MoleculeProcess`] record (the inline slot
        `molecule.process`, added by `delib-20260426-1bcd` to collapse
        the legacy tri-witness `assigned_worker` / `current_molecule`
        / `session_name`). Presence of the slot means "the pilot
        believes a live worker is bound"; absence means no worker.

        Tmux is **not** consulted directly — tmux sessions live in a
        host-global namespace and re-introducing it as a witness would
        bring back the phantom-worker disagreement the fold-in fixed.

        ## Required scope

        `cosmon:worker:read` (distinct from `cosmon:molecule:read` —
        the worker surface exposes session-level facts such as the
        tmux session name, PID, and start instant, which a tenant
        scoped only for molecule state should not be granted).

        ## Empty noyau

        When no molecule carries a live process record, the response
        is `{ workers: [], count: 0 }` (200, not 204). The shape stays
        stable for tenant code that always parses the array.

        ## Pagination + ordering

        V1 returns the full list. Entries are sorted by `started_at`
        ascending so a paginating client sees a deterministic ordering
        across calls. The count is small in practice (a noyau rarely
        runs more than a few dozen concurrent workers).
      operationId: listWorkers
      tags: [workers]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: |
            Active workers in the caller's noyau. Empty array on a
            freshly-provisioned noyau or when no molecule carries a
            live process.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkersResponse"
        "401":
          description: |
            Missing / malformed `Authorization` header, JWT signature
            invalid, expired, unknown issuer, or audience mismatch.
            One stable label per reject reason; the wire body never
            leaks `sub` / nucleon / tenant identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: |
            JWT lacks `cosmon:worker:read`. The route does NOT fall
            back to `cosmon:molecule:read` — observability over the
            worker surface is a strictly distinct scope by design.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: |
            The admission rate-limit already exceeded — the wire
            response still carries `X-RateLimit-*` headers + a
            `Retry-After` so the client knows when to retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: |
            The per-tenant store could not be read. Transient — the
            adapter does not crash, the next call may succeed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC-issued JWT signed with `RS256` or `ES256` (clause (a)).
        The token's `kid` MUST resolve in the operator's pinned JWKS
        (`<state_dir>/security/jwks/`). `none`, `HS256`, `EdDSA`,
        `RS384`, `RS512` are rejected at parse time.
  schemas:
...
    VitalsEnvelope:
      type: object
      required: [request_id, vitals]
      properties:
        request_id:
          type: string
        vitals:
          $ref: "#/components/schemas/VitalsView"
    VitalsView:
      type: object
      required: [molecules, counts, backlog, observed_at]
      properties:
        molecules:
          type: array
          description: Exactly one row per non-terminal molecule.
          items:
            $ref: "#/components/schemas/VitalsRow"
        counts:
          $ref: "#/components/schemas/VitalsCounts"
        backlog:
          $ref: "#/components/schemas/VitalsBacklog"
        observed_at:
          type: string
          format: date-time
    VitalsRow:
      type: object
      required:
        [id, status, current_step, total_steps, health,
         awaiting_operator, created_at, updated_at]
      properties:
        id:
          type: string
        status:
          type: string
        current_step:
          type: integer
          minimum: 0
        total_steps:
          type: integer
          minimum: 0
        worker:
          type: string
        health:
          type: string
          enum: [unassigned, live, orphaned, unknown]
        awaiting_operator:
          type: boolean
          description: True only for the declared `temp:awaiting-op` tag.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    VitalsCounts:
      type: object
      required: [molecules, live, orphaned, unknown, unassigned, awaiting_operator]
      properties:
        molecules: { type: integer, minimum: 0 }
        live: { type: integer, minimum: 0 }
        orphaned: { type: integer, minimum: 0 }
        unknown: { type: integer, minimum: 0 }
        unassigned: { type: integer, minimum: 0 }
        awaiting_operator: { type: integer, minimum: 0 }
    VitalsBacklog:
      type: object
      required: [count, stale, oldest_age_seconds, oldest_id, leases_excluded]
      properties:
        count: { type: integer, minimum: 0 }
        stale: { type: integer, minimum: 0 }
        oldest_age_seconds:
          type: [integer, "null"]
          minimum: 0
        oldest_id:
          type: [string, "null"]
        leases_excluded: { type: integer, minimum: 0 }
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: |
            Coarse-grained error label.

            **Open by contract.** The server may emit new labels in
            future versions without an API bump. Callers MUST treat
            any unknown value as semantically equivalent to
            `internal` (i.e. retry-on-idempotent, surface-as-opaque).
            Known values at v1.0.0 include `unauthorized`,
            `forbidden`, `not_found`, `invalid_filter`,
            `invalid_json_body`, `missing_reason`, `empty_reason`,
            `empty_tag_request`, `terminal_status`, `invalid_status`,
            `rate_limited`, `store_unavailable`, `endpoint_removed`,
            `internal` — but this list is *informative*, not
            normative. Mirror of the lenient `#[non_exhaustive]` +
            `#[serde(other)] Unknown(String)` discipline already
            inscribed for `Status` / `Kind` / `MoleculeEventKind`.

            The wire body never carries `sub`, `nucleon_id`, or any
            tenant-identifying field.
        request_id:
          type: string
          description: |
            Optional correlation id; present on errors that
            materialise an audit envelope.
        retry_after_seconds:
          type: integer
          format: int64
          minimum: 0
          description: |
            Present only when `error` is `rate_limited`. The number of
            whole seconds to wait before retrying, rounded up from the
            admission boundary's precise delay. The `Retry-After` response
            header carries this same decimal value; clients may use either
            representation and MUST NOT retry a non-idempotent mutation
            without an idempotency key.
...
    WorkersResponse:
      type: object
      required: [request_id, workers, count]
      description: |
        Response body for `GET /v1/workers` (`task-20260523-f82b`,
        gap report ae3d workflow §e priority 5). Lists the active
        workers in the caller's noyau.
      properties:
        request_id:
          type: string
          description: |
            Audit-correlation id for the inbox-materialised admission
            row of this `/v1/workers` call.
        workers:
          type: array
          description: |
            Active workers, sorted by `started_at` ascending so a
            paginating client sees a deterministic enumeration.
            Empty when no molecule carries a live process record.
          items:
            $ref: "#/components/schemas/WorkerEntry"
        count:
          type: integer
          format: int64
          minimum: 0
          description: Number of active workers (workers.length).
    WorkerEntry:
      type: object
      required: [molecule_id, session_name, started_at, tmux_session]
      description: |
        One worker entry inside [`WorkersResponse`]. Stable
        additive-only wire shape: adding a field is allowed; renaming
        or removing one is a §8p break.
      properties:
        molecule_id:
          type: string
          description: |
            Molecule the worker is bound to (the inline
            `MoleculeProcess` slot on the molecule's state).
        session_name:
          type: string
          description: |
            Worker identity stamped at `cs tackle` time — the same
            string the operator sees in `cs status` rows. Currently
            equal to the tmux session name; the two fields are kept
            distinct so they can evolve independently.
        started_at:
          type: string
          format: date-time
          description: |
            ISO-8601 / RFC 3339 instant when the process record was
            created (typically by `cs tackle`).
        pid:
          type: integer
          format: int64
          minimum: 0
          nullable: true
          description: |
            Operating-system PID when the transport backend surfaced
            one. `null` for backends that do not expose PIDs (legacy
            or in-process adapters). The field is always present so
            clients can branch on its absence rather than on a
            missing key.
        tmux_session:
          type: string
          description: |
            Tmux session owning the worker process — same string that
            `tmux -L <socket> attach -t <session>` would target.
...